+++ sphinx/source/changelog.rst
@@ -26,6 +26,8 @@ Changelog
   transparent colour)
 * Added the :ref:`refs` directive (for specifying the addresses of routines
   that indirectly jump to an entry point)
+* Added support for replacement fields in the integer parameters of the
+  :ref:`FOR` macro
 * Added the ``--page`` option to :ref:`snapinfo.py` (for specifying the page of
   a 128K snapshot to map to 49152-65535)